+++ CHANGELOG.md
@@ -227,6 +227,11 @@
   **2.13.0**
   - fixed: skip-backup
   - added: skip-delete-google
+  **2.13.1**
+  - fixed fixed: skip-backup
+  - fixed: enter upload
+  **2.13.2**
+  - updated: google mimetypes
 
 ----------------------------------------